requirements.txt
Jinja2==3.1.2
requests==2.31.0
pylint==2.17.4
pytest-mock==3.11.1
